improve PlayerListenerCommon
